Fixing issues.

+++ Kernel/include/siberix/device/connectivity.h
@@ -24,12 +24,12 @@ private:
     LinkedList<ConnectivityProvider&> m_providers;
     /// @brief list of connected devices, those unrecognizable devices will be
     /// categoried into 'unknown devices'
-    LinkedList<Device&>               m_deviceList[DeviceType::Unknown + 1];
+    LinkedList<Device&> m_deviceList[static_cast<u64>(DeviceType::Unknown) + 1];
     /// @brief the amount of connected devices
-    u32                               m_connectedDeviceCount;
+    u32                 m_connectedDeviceCount;
     /// @brief auto connect and load new but unneccssary devices (bluetooth,
     /// external hard drive, etc.) when they are available if set to true
-    bool                              m_isAutoConnect;
+    bool                m_isAutoConnect;
 
 protected:
     friend Device;
